Consolidate test helpers
